app/requirements.txt
psycopg2-binary
numpy
SQLAlchemy
ipython
